+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/Rules/ChangeConnectionStatus/ConnectionStateMustBeConnectedOrDisconnectedError.cs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules.Disconnect
+namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules.ChangeConnectionStatus
 {
     public class ConnectionStateMustBeConnectedOrDisconnectedError : ValidationError
     {
