+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/Rules/ChangeConnectionStatus/PhysicalStateMustBeConnectedOrDisconnectedRule.cs
@@ -15,7 +15,7 @@
 using System;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules.Disconnect
+namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules.ChangeConnectionStatus
 {
     public class PhysicalStateMustBeConnectedOrDisconnectedRule : IBusinessRule
     {
