+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/Errors/Converters/Disconnect/ConnectionStateMustBeConnectedOrDisconnectedErrorConverter.cs
@@ -14,7 +14,7 @@
 
 using System;
 using Energinet.DataHub.MeteringPoints.Application.EDI;
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules.Disconnect;
+using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules.ChangeConnectionStatus;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors.Converters.Disconnect
 {
